requirements.txt
python-dotenv 
streamlit
openai 
azure-cli 
azure-ai-documentintelligence 
azure-storage-blob
azure-search-documents
tenacity 
numpy